Add files via upload
The red curve is the error rate for the Setosa class, the green and blue curves above are for Versicolor and Virginica while the black curve is the Out-of-Bag error rate. 

We reamrk that this plot shows how the OOB error rate (proportion of misclassifications) for each of the three species changes with the size of the forest (the number of trees). With few trees the error rate is high, but as more trees are added you can see the error rate decrease and eventually flatten out.